+++ Backend/src/main/java/co/edu/uniquindio/proyecto/dto/ProFuncioDto/EditarFuncioDto.java
@@ -2,6 +2,8 @@ package co.edu.uniquindio.proyecto.dto.ProFuncioDto;
 
 import jakarta.validation.constraints.*;
 
+import java.math.BigInteger;
+
 public record EditarFuncioDto(
         @NotNull Integer id_funcio,
         @NotNull Integer id_tipdoc,
@@ -13,7 +15,7 @@ public record EditarFuncioDto(
         @NotBlank @Size(max = 25) String nm_func2,
         @NotBlank @Size(max = 25) String ap_func1,
         @NotBlank @Size(max = 25) String ap_func2,
-        @NotNull Integer no_funcio,
+        @NotNull BigInteger no_funcio,
         @NotBlank @Size(max = 100) String ce_funcio
 ) { }
